use role terminology throughout
