+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/RouteConfig.kt
@@ -30,6 +30,6 @@ object WebRouteConfig : RouteConfig
                     WebReportRouteConfig.endpoints +
                     WebVersionRouteConfig.endpoints +
                     WebUserRouteConfig.endpoints +
-                    WebUserGroupRouteConfig.endpoints + metricsEndpoint
+                    WebRoleRouteConfig.endpoints + metricsEndpoint
 }
