+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/app_start/routing/web/WebUserRouteConfig.kt
@@ -3,6 +3,7 @@ package org.vaccineimpact.orderlyweb.app_start.routing.web
 import org.vaccineimpact.orderlyweb.*
 import org.vaccineimpact.orderlyweb.app_start.RouteConfig
 import org.vaccineimpact.orderlyweb.controllers.web.UserController
+import org.vaccineimpact.orderlyweb.controllers.web.UserGroupController
 import spark.route.HttpMethod
 
 object WebUserRouteConfig : RouteConfig
@@ -10,10 +11,15 @@ object WebUserRouteConfig : RouteConfig
     private val usersManage = setOf("*/users.manage")
     override val endpoints = listOf(
             WebEndpoint("/users/report-readers/:report/",
-                UserController::class, "getScopedReportReaders",
+                    UserController::class, "getScopedReportReaders",
                     contentType = ContentTypes.json)
-                .json()
-                .transform()
-                .secure(usersManage)
+                    .json()
+                    .transform()
+                    .secure(usersManage),
+            WebEndpoint("/user-groups/:user-group-id/actions/associate-permission/",
+                    UserGroupController::class, "associatePermission",
+                    method = HttpMethod.post)
+                    .json()
+                    .secure(usersManage)
     )
 }
