+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/controllers/web/UserController.kt
@@ -4,18 +4,18 @@ import org.vaccineimpact.orderlyweb.ActionContext
 import org.vaccineimpact.orderlyweb.controllers.Controller
 import org.vaccineimpact.orderlyweb.db.OrderlyUserRepository
 import org.vaccineimpact.orderlyweb.db.UserRepository
-import org.vaccineimpact.orderlyweb.viewmodels.ReportReaderViewModel
+import org.vaccineimpact.orderlyweb.viewmodels.UserViewModel
 
 class UserController(context: ActionContext,
                      private val userRepo: UserRepository) : Controller(context)
 {
     constructor(context: ActionContext) : this(context, OrderlyUserRepository())
 
-    fun getScopedReportReaders(): List<ReportReaderViewModel>
+    fun getScopedReportReaders(): List<UserViewModel>
     {
         val report = report()
-        val users = userRepo.getScopedIndividualReportReaders(report)
-        return users.map { ReportReaderViewModel.build(it) }
+        val users = userRepo.getScopedReportReaderUsers(report)
+        return users.map { UserViewModel.build(it) }
                 .sortedBy { it.displayName.toLowerCase() }
     }
