+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/models/permissions/Role.kt
@@ -2,4 +2,4 @@ package org.vaccineimpact.orderlyweb.models.permissions
 
 import org.vaccineimpact.orderlyweb.models.User
 
-data class UserGroup(val name: String, val members: List<User>)
+data class Role(val name: String, val members: List<User>)
