+++ src/app/src/main/kotlin/org/vaccineimpact/orderlyweb/viewmodels/RoleViewModel.kt
@@ -1,0 +1,16 @@
+package org.vaccineimpact.orderlyweb.viewmodels
+
+import org.vaccineimpact.orderlyweb.models.permissions.Role
+
+data class RoleViewModel(val name: String, val members: List<UserViewModel>)
+{
+    companion object
+    {
+        fun build(role: Role): RoleViewModel
+        {
+            return RoleViewModel(role.name, role.members.map {
+                UserViewModel.build(it)
+            }.sortedBy { it.displayName })
+        }
+    }
+}
