+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/integration_tests/tests/web/UsersTests.kt
@@ -30,51 +30,4 @@ class UsersTests : IntegrationTest()
                 "name" to "users.manage"))
     }
 
-    @Test
-    fun `only user managers can get global report reading groups`()
-    {
-        val url = "/user-groups/report-readers/"
-
-        assertWebUrlSecured(url, setOf(ReifiedPermission("users.manage", Scope.Global())),
-                contentType = ContentTypes.json)
-    }
-
-    @Test
-    fun `can get global report reading groups`()
-    {
-        createGroup("Funder", ReifiedPermission("reports.read", Scope.Global()))
-        addMembers("Funder", "funder.a@example.com", "funder.b@example.com")
-
-        val url = "/user-groups/report-readers/"
-        val response = webRequestHelper.loginWithMontaguAndMakeRequest(url,
-                setOf(ReifiedPermission("users.manage", Scope.Global())),
-                ContentTypes.json)
-
-        JSONValidator.validateAgainstSchema(response.text, "UserGroups")
-    }
-
-
-    @Test
-    fun `can get scoped report reading groups`()
-    {
-        createGroup("Funder", ReifiedPermission("reports.read", Scope.Specific("report", "minimal")))
-        addMembers("Funder", "funder.a@example.com", "funder.b@example.com")
-
-        val url = "/user-groups/report-readers/minimal/"
-        val response = webRequestHelper.loginWithMontaguAndMakeRequest(url,
-                setOf(ReifiedPermission("users.manage", Scope.Global()),
-                        ReifiedPermission("reports.read", Scope.Global())),
-                ContentTypes.json)
-
-        JSONValidator.validateAgainstSchema(response.text, "UserGroups")
-    }
-
-    @Test
-    fun `only user managers can get all role names`()
-    {
-        val url = "/typeahead/roles/"
-
-        assertWebUrlSecured(url, setOf(ReifiedPermission("users.manage", Scope.Global())),
-                contentType = ContentTypes.json)
-    }
 }
