+++ src/app/src/test/kotlin/org/vaccineimpact/orderlyweb/tests/unit_tests/controllers/web/UserControllerTests.kt
@@ -32,7 +32,7 @@ class UserControllerTests : TeamcityTests()
         )
 
         val repo = mock<UserRepository> {
-            on { this.getScopedIndividualReportReaders("r1") } doReturn (reportReaders)
+            on { this.getScopedReportReaderUsers("r1") } doReturn (reportReaders)
         }
         val sut = UserController(actionContext, repo)
         val result = sut.getScopedReportReaders()
@@ -72,7 +72,7 @@ class UserControllerTests : TeamcityTests()
                         "r4@email.com"))
 
         val repo = mock<UserRepository> {
-            on { this.getScopedIndividualReportReaders("r1") } doReturn (reportReaders)
+            on { this.getScopedReportReaderUsers("r1") } doReturn (reportReaders)
         }
         val sut = UserController(actionContext, repo)
         val result = sut.getScopedReportReaders()
